Add example of analyzing soundness of Send/Sync
Add an example of thinking through whether it is sound to implement Send
+ Sync for a custom type that wraps a raw pointer.

I read the existing docs and was confused about whether I could
implement Send + Sync for a type I wrote that wraps a c-style array.
Kixiron, InfernoDeity, Talchas, and HeroicKatora on #black-magic
helped me understand Send and Sync better. This example is based on the
advice they gave me. I've made lots of changes, so any errors are
probably mine.