Two lifetime clarification
In the first change "this" and "that" is slightly ambiguous, at least for a
first introduction. Actually, I'm not 100% sure I got it right, but at least, if
I got it wrong, it proves my point it needed clarification

The second change is simply because "x has to live (for 'b to be printed)"
seemeds to mean that "'b" will be printed. I hope it makes clear that the actual
parsing is "(x has to live for 'b) in order to be printed"