Rephrase improperly reduced borrows introduction
The issue with the current version of this paragraph is that it refers to code
that will occur later. It can only be understood once we know what is borrowed
and when, which only becomes clear when the error message is shown.

With this new phrasing, I hope that the introduction prepares the reader to what
they'll find out.

Another solution would be to imitate the first example of this page. Introduce
the code with just "Given the following code:" and then comment why it
failed.