Introducing init/uninit before its use
In this case, I believe the meaning is intuitive enough that the abbreviation
makes sens. However, I still prefer to introduce it explicitly. In particular
because, in drop-flags, I misintepreted "x was uninit". I thought that "uninit"
was a verb, so that something did uninit x. It is only after some thought that I
understood that it was not a passive sentence but a descriptive one, and that
"uninit" should be understood as an adjective.