Include elided lifetime in sub-typing example
One step required to understand the argumenation is that the lifetime of
the `&mut T` is different from any lifetime that may be part of a
reference forming the type T. By mentioning the elided lifetime
explictly, the paragraph should be easier to understand.

Minor other changes are highlighting types with backticks.