Rename generic type parameter in Subtyping chapter
`T` was used to refer both to the generic type parameter of the function
used as an example, and to the type that this generic type parameter
would be a reference to. To make things clearer, the generic type
parameter is renamed to `R`, since it is intended to be a reference in
the examples.